+++ lib/Features/Auth/reset_password/presentation/reset_password_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:study_mate/Features/Auth/login/presentation/widgets/reset_password_body.dart';
+import 'package:study_mate/Features/Auth/reset_password/presentation/reset_password_body.dart';
 
 class ResetPasswordScreen extends StatelessWidget {
   const ResetPasswordScreen({super.key});
